Modularized navbar
